PostView2: Added colormap tab to view settings. Users can now edit existing and add new colormaps.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/PostView2@9483 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ PostViewLib/DataMap.h
@@ -14,7 +14,7 @@ public:
 	DataMap(void) { m_pmesh = 0; }
 	~DataMap(void) {}
 
-	int States() { return m_Data.size(); }
+	int States() { return (int)m_Data.size(); }
 	vector<T>& State(int n) { return m_Data[n]; }
 	int GetTag(int n) { return m_tag[n]; }
 	void SetTag(int n, int ntag) { m_tag[n] = ntag; }
